UPDATE: move classes to correct layer
